+++ streams/vox/vox_loader.h
@@ -1,6 +1,7 @@
 #ifndef VOX_LOADER_H
 #define VOX_LOADER_H
 
+#include "../../storage/voxel_buffer_gd.h"
 #include "../../util/godot/classes/ref_counted.h"
 
 namespace zylann::voxel {
@@ -17,7 +18,8 @@ class VoxelVoxLoader : public RefCounted {
 
 public:
 	// TODO GDX: Can't bind functions returning a `godot::Error` enum
-	int /*Error*/ load_from_file(String fpath, Ref<gd::VoxelBuffer> p_voxels, Ref<VoxelColorPalette> palette);
+	int /*Error*/ load_from_file(String fpath, Ref<gd::VoxelBuffer> p_voxels, Ref<VoxelColorPalette> palette,
+			gd::VoxelBuffer::ChannelId dst_channel);
 	// TODO Have chunked loading for better memory usage
 	// TODO Saving