+++ thirdparty/fast_noise_2/include/FastNoise/FastNoise_Config.h
@@ -16,7 +16,8 @@ namespace FastNoise
         FastSIMD::Level_SSE2   |
         FastSIMD::Level_SSE41  |
         FastSIMD::Level_AVX2   |
-        FastSIMD::Level_AVX512 ;
+        FastSIMD::Level_AVX512 |
+        FastSIMD::Level_NEON   ;
     
     class Generator;
     struct Metadata;